18.5 .env.sample was updated.
Status_codes in autho.py were changed, as it was requested.

declarative_base was relocated to base.py, models.py was deleted and replaced with model directory which contains separated models.

celery_app.py was not changed, since it work in current iteration. Other types of implementations cause errors.

As it was asked, return types were added for all of the CRUDs and services.

health_check.py was separated to a new router file, health_check_router.py.

One error handling was deleted from user_service.py, as it was not needed. According test for error_handling was deleted too.

New __init__.py was added. Celery throws an error connected to relationships without that file.